code cleanup and commenting brought about when documenting for jhu report
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk/moses@886 1f5c12ca-751b-0410-a591-d2e778427230

+++ src/LanguageModelIRST.cpp
@@ -153,7 +153,7 @@ float LanguageModelIRST::GetValue(const vector<const Word*> &contextFactor, Stat
 		*len = 0;	
 	}
 
-	return TransformIRSTScore(m_lmtb->clprob(*m_lmtb_ng));
+	return TransformIRSTScore((float) m_lmtb->clprob(*m_lmtb_ng));
 }
